Use availability check for iOS 15 styling funcs
