+++ DuckDuckGo/NetworkProtectionStatusView.swift
@@ -43,7 +43,7 @@ struct NetworkProtectionStatusView: View {
         .animation(.default, value: statusModel.shouldShowError)
         .padding(.top, statusModel.error == nil ? 0 : -20)
         .animation(.default, value: statusModel.shouldShowConnectionDetails)
-        .applyInsetGroupedListStyleForiOS15AndOver()
+        .applyInsetGroupedListStyle()
         .navigationTitle(UserText.netPNavTitle)
     }
 
@@ -75,7 +75,7 @@ struct NetworkProtectionStatusView: View {
         } header: {
             header()
         }
-        .increaseHeaderProminenceForiOS15AndOver()
+        .increaseHeaderProminence()
     }
 
     @ViewBuilder
